% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read20NewsGroups.R
\name{get_20newsgroup_data}
\alias{get_20newsgroup_data}
\title{Read 20 Newsgroup data}
\usage{
get_20newsgroup_data(g1 = 10, g2 = 14, rate = 1,
  remove_stopwords = FALSE, min_threshold = 0, data.file = NA,
  labels.file = NA)
}
\arguments{
\item{g1, g2}{Integers. 20 newsgroups to compare. Default: g1=10 (rec.sport.baseball), g2=14 (sci.med)}

\item{rate}{Decimal number in the range of [0,1]. Sampling rate at which to sample records (without replacement). Default=1.}

\item{remove_stopwords}{Boolean. TRUE = remove stopwords. Default=FALSE.}

\item{min_threshold}{Integer value. Words having count < threshold will be removed. remove words that have counts less than or equal to threshold (looking across entire group of docs)}

\item{data.file}{File path to the file containing the 20 newsgroup data.}

\item{labels.file}{File path to the file containing the 20 newsgroup labels.}
}
\value{
Two tf vectors, one for each group.
}
\description{
Read 20 Newsgroup data
}
\examples{
#Generate two vectors from the same distribution:
data <- genMultinomialData(sample_size=1)

}
