% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtest.R
\docType{package}
\name{hddtest}
\alias{hddtest}
\alias{hddtest-package}
\title{hddtest: A package for testing two multinomial datasets.}
\description{
The hdd test package provides the following functions:
multinom.test, genMultinomialData, etc.
}
\section{hddtest functions}{

}

