% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinom.neighborhood.test}
\alias{multinom.neighborhood.test}
\title{Perform the neighborhood test for multinom.test}
\usage{
multinom.neighborhood.test(x, y = NULL, delta = NULL)
}
\arguments{
\item{x, y}{Integer vectors (or matrices/dataframes containing multiple
integer vector observations as rows). \code{x} and \code{y} must be the
same type and dimension. If \code{x} and \code{y} are matrices (or
dataframes), the \eqn{i^th} row of \code{x} will be tested against the
\eqn{i^th} row of \code{y} for all \eqn{i} in 1..\code{nrow(x)}.
Alternatively, \code{x} can be a list of two vectors, matrices, or
dataframes to be compared. In this case, \code{y} is NULL by default.}

\item{delta}{A number (or vector) greater than 0.}
}
\value{
The \code{statistic} from \code{multinom.test}, T, and its
associated \code{p_delta}, where \eqn{p_delta=pnorm(T - delta)}. TODO: Why subtract from 1 in code?
If \code{x} and \code{y} are either matrices or dataframes, a
\code{statistic} and \code{p-value} will be returned for each row.
}
\description{
In testing the equality of parameters from two populations
(as in \code{multinom.test}),
it frequenly happens that the null hypothesis is rejected even though the estimates
of effect sizes are close to each other. However, these differences are so small
that the parameters may not be considered different in practice. A neighborhood test
is useful in this situation.
}
\details{
Peforms the test for two multinomial vectors
testing \eqn{H_0:} the underlying multinomial probability vectors are the
same vs. \eqn{H_1:} they are different.
}
\examples{
#Generate two vectors from the same distribution:
data <- genMultinomialData(sample_size=1)

#Perform test (the following calls are equivalent):
delta=60
multinom.neighborhood.test(x=data[[1]],y=data[[2]],delta=delta)
multinom.neighborhood.test(data,delta=delta)
library(magrittr)
data \%>\% multinom.neighborhood.test(delta=delta)

#Generate 1000 vectors from each of two different distributions:
data <- genMultinomialData(null_hyp=FALSE,sample_size=1000)

#Perform test (compare the ith row of x to the ith row of y for all rows):
result <- multinom.neighborhood.test(x=data[[1]],y=data[[2]],delta=delta)

#Return power of test at the alpha=0.05 level:
alpha <- 0.05
mean(result$pvalue<alpha)
}
\seealso{
Amanda Plunkett & Junyong Park (2018) \emph{Two-Sample Test for Sparse High
Dimensional Multinomial Distributions}, TEST,
\url{https://doi.org/10.1007/s11749-018-0600-8}
}
