% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinom.test}
\alias{multinom.test}
\title{Test two multinomial datasets}
\usage{
multinom.test(x, y)
}
\arguments{
\item{x, y}{A list of two matrices}
}
\value{
The \code{statistic} and its associated \code{p-value}
}
\description{
TODO: Add a longer description here.
}
\examples{
data <- genMultinomialData(null_hyp=TRUE)
multinom.test(x=data[[1]],y=data[[2]])
}
