% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{mvbinary.test}
\alias{mvbinary.test}
\title{Test two multivariate binary datasets}
\usage{
mvbinary.test(x, y = NULL, numPerms = 5000)
}
\arguments{
\item{x, y}{Matrices (or dataframes) containing multiple
integer vector observations as rows. \code{x} and \code{y} must be the
same type and dimension. Alternatively, \code{x} can be a list of two
matrices (or dataframes) to be compared. In this case, \code{y} is NULL
by default.}

\item{numPerms}{Number of permutations to use to calculate the p-value.
Default value is 5000.}
}
\value{
The \code{statistic} and its associated \code{p-value}.
}
\description{
Peforms the test for two binary vectors
testing \eqn{H_0:} the underlying probability vectors are the
same vs. \eqn{H_1:} they are different.
}
\details{
Statistic:
\eqn{T = \sum_{j=1}^d D_j^2 I( |Dj| \ge \delta(d))}
where \eqn{d} is the dimension.
\eqn{Dj = (\hat{p}_{1j} − \hat{p}_{2j} )/sqrt{ \hat{p}_j (1 − \hat{p}_j )(1/n1 + 1/n2) } }
where \eqn{\hat{p}_{cj}} is the estimate of \eqn{p_{cj}} for the \eqn{c^{th}}
group calculated by  \eqn{x_{cj}}. Let \eqn{\hat{p}_j} be the pooled estimate
for the \eqn{j^{th}} variable. Additionally,
\eqn{\delta(d) = \sqrt{2 log (a_d d)}} where \eqn{a_d = (log d)^{-2}}.

P-value: Calculated using the permutation method.

Note: As described in the reference below, this method does not perform
well on highly correlated variables. See the reference for more details.
}
\examples{
#Binarize the twoNewsGroups dataset:
binData <- list(twoNewsGroups[[1]]>0,twoNewsGroups[[2]]>0)
names(binData) <- names(twoNewsGroups)

#Perform the test:
mvbinary.test(binData)

#The following are equivalent to the previous call:
mvbinary.test(binData[[1]],binData[[2]])
require(magrittr)
binData \%>\% mvbinary.test
}
\seealso{
Amanda Plunkett & Junyong Park (2017) \emph{Two-sample Tests for Sparse
High-Dimensional Binary Data}, Communications in Statistics - Theory and
Methods, 46:22, 11181-11193
}
