% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{genMultinomialData}
\alias{genMultinomialData}
\title{Generate multinomial data}
\usage{
genMultinomialData(p = NULL, null_hyp = TRUE, k = 2000, n = c(8000,
  8000), sample_size = c(30, 30), expID = 1, alpha = 0.45, numzero = 50,
  beta = 0.25)
}
\arguments{
\item{p}{An optional 2 by k matrix specifying the probabilities of the k categories for each of the two groups.
If defined the rest of the function parameters will not be used. Default value is NULL.}

\item{null_hyp}{logical; if TRUE, generate data using the same distribution.}

\item{k}{dimension (number of categories). Default 2000.}

\item{n}{Vector of length 2 specifying the parameter of the multinomial distribution used to specify the
total number of objects that are put into k bins in the typical multinomial experiment.}

\item{sample_size}{A vector of length 2 specifying the number of multinomial vectors to generate}

\item{expID}{Experiment number 1-6}

\item{alpha}{=0.45}

\item{numzero}{=50}

\item{beta}{=0.25}
}
\value{
A list containing two dataframes each having dimension sample_size by k.
}
\description{
TODO: Add a longer description here.
}
\examples{
#Generate data:
X <- genMultinomialData(null_hyp=TRUE)
#Look at the first 10 rows and columns of the first matrix:
X[[1]][1:10,1:10]
}
