% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{genMultinomialData}
\alias{genMultinomialData}
\title{Generate multinomial data}
\usage{
genMultinomialData(p = NULL, null_hyp = TRUE, k = 2000, n = c(8000,
  8000), sample_size = 30, expID = 5, alpha = 0.45, numzero = 50,
  beta = 0.25)
}
\arguments{
\item{p}{An optional 2 by k matrix specifying the probabilities of the k categories for each of the two groups.
Each row of p must sum to 1.
If defined the rest of the function parameters will not be used. Default value is NULL.}

\item{null_hyp}{logical; if TRUE, generate data using the same distribution. Default value is TRUE.}

\item{k}{integer representing dimension (number of categories). Default 2000.}

\item{n}{Vector of length 2 specifying the parameter of each multinomial distribution used to specify the
total number of objects that are put into k bins in the typical multinomial experiment.}

\item{sample_size}{An integer specifying the number of random vectors to generate for each of the two groups}

\item{expID}{Experiment number 1-6}

\item{alpha}{Number between 0 and 1. Default is 0.45. Used for experiments 5 and 6.}

\item{numzero}{Integer between 1 and k-1. Default is 50. Used for experiments 1-3.}

\item{beta}{Default is 0.25. TODO}
}
\value{
A list containing two matrices each having dimension sample_size by k.
}
\description{
Generate two sets of multinomially distributed vectors using rmultinom. Useful for hypothesis testing simulations.
Six different experiments with different probability vectors are available in addition to user-specified probability vector \code{p}:
\itemize{
\item Experiment1: \eqn{p_{1i} = 1/k, p_{2i} = 0} for \eqn{i \in [1,b]} and \eqn{p_{2i} = 1/(k − b)}  for \eqn{i > b}
\item Experiment2: \eqn{p_{1i} = 1/k, p_{2i} = 0} for \eqn{i \in [1,b]} and \eqn{p_{2,b+1}= \sum_{1}^{b+1} p_{2i} }
\item Experiment3: TODO
\item Experiment4: TODO
\item Experiment5: \eqn{p_{1i} = \frac{1/i^\alpha}{\sum_1^k 1/i^\alpha}}
\item Experiment6: TODO
}
}
\examples{
#Generate data:
X <- genMultinomialData(null_hyp=TRUE)

#Dimension of generated datasets:
dim(X[[1]])
dim(X[[2]])

Summary of first three columns of the first dataset:
summary(X[[1]][,1:3])
}
