% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{genMultinomialData}
\alias{genMultinomialData}
\title{Generate multinomial data}
\usage{
genMultinomialData(
  null_hyp = TRUE,
  p = NULL,
  k = 2000,
  n = c(8000, 8000),
  sample_size = 30,
  expID = 1,
  alpha = 0.45,
  m = 1000,
  numzero = 50,
  ...
)
}
\arguments{
\item{null_hyp}{logical; if TRUE, generate data using the same distribution.
Default value is TRUE.}

\item{p}{An optional 2 by \eqn{k} matrix specifying the probabilities of the
\eqn{k} categories for each of the two groups. Each row of \code{p} must sum
to 1. If defined, all remaining parameters in the function definition are
ignored. Default value is NULL.}

\item{k}{integer representing dimension (number of categories). Default 2000.}

\item{n}{Vector of length 2 specifying the parameter of each multinomial
distribution used to define the total number of objects that are put into
\eqn{k} bins in the typical multinomial experiment.}

\item{sample_size}{integer specifying the number of random vectors to generate
for each of the two groups.}

\item{expID}{Experiment number 1-3. Default is 1.}

\item{alpha}{Number between 0 and 1. Used for experiment 1. Default is 0.45.}

\item{m}{integer between 2 and \eqn{k}. Used in experiment 1 for the
alternative hypothesis. Default is 1000.}

\item{numzero}{integer between 1 and \eqn{k}-1. Used in experiments 2 and 3
for the alternative hypothesis. Default is 50.}

\item{...}{Additional parameters.}
}
\value{
A list containing two matrices each having dimension
\code{sample_size} by \eqn{k}.
}
\description{
Generate two sets of multinomially distributed vectors using
\code{rmultinom}. Useful for hypothesis testing simulations. Three different
experiments with different probability vectors (of length \eqn{k}) are
available in addition to user-specified probability vector \code{p}:
\itemize{
\item Experiment 1: \eqn{p_{1i} = \frac{1/i^\alpha}{\sum_1^k 1/i^\alpha}}.
      When the \code{null_hyp} parameter is FALSE, the probability vector
      for the 2nd group is generated by switching the position of 1st and
      \eqn{m^th} entries.
\item Experiment 2: \eqn{p_{1i} = 1/k}. When the \code{null_hyp} parameter is FALSE,
      \eqn{p_{2i} = 0} for \eqn{i \in 1...b} and
      \eqn{p_{2,b+1}= \sum_{1}^{b+1} p_{1i} = (b+1)/k }.
\item Experiment 3: \eqn{p_{1i} = 1/k}. When the \code{null_hyp} parameter is FALSE,
      \eqn{p_{2i} = 0} for \eqn{i \in 1...b} and \eqn{p_{2i} = 1/(k − b)} for
      \eqn{i > b}.
}
}
\examples{
#Generate data when the null hypothesis is FALSE:
X <- genMultinomialData(FALSE)

#Dimension of the two generated datasets:
lapply(X, dim)

#Proportion of entries less than 5 in the first dataset:
sum(X[[1]]<5)/(nrow(X[[1]])*ncol(X[[1]]))

}
