% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddtest.R
\docType{package}
\name{hddtest}
\alias{hddtest}
\alias{hddtest-package}
\title{hddtest: Two-sample hypothesis testing for high dimensional discrete disributions}
\description{
The hdd test package provides functionality for two-sample hypothesis testing of high dimensional discrete data. Currently, testing of multinomial data is supported.
}
\section{hddtest functions}{

multinom.test,
genMultinomialData,
}

\section{Authors}{

Amanda Plunkett, Junyong Park
}

\section{References}{

\itemize{
\item Amanda Plunkett & Junyong Park (2018) \emph{Two-Sample Test for Sparse High Dimensional Multinomial Distributions}, TEST, \url{https://doi.org/10.1007/s11749-018-0600-8}
\item Amanda Plunkett & Junyong Park (2017) \emph{Two-sample Tests for Sparse High-Dimensional Binary Data}, Communications in Statistics - Theory and Methods, 46:22, 11181-11193
}
}

