% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinom.test}
\alias{multinom.test}
\title{Test two multinomial datasets.}
\usage{
multinom.test(data)
}
\arguments{
\item{data}{A list of two matrices}
}
\value{
The \code{statistic} and its associated \code{p-value}
}
\description{
Test two multinomial datasets.
}
\examples{

}
