% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinom.test}
\alias{multinom.test}
\title{Test two multinomial datasets}
\usage{
multinom.test(x, y)
}
\arguments{
\item{x, y}{Integer vectors (or matrices/dataframes containing multiple integer vector observations as rows).
\code{x} and \code{y} must be the same type and dimension. If \code{x} and \code{y} are matrices (or dataframes), the ith row
of \code{x} will be tested against the \eqn{i^th} row of \code{y} for all \eqn{i} in 1..nrow(x).}
}
\value{
The \code{statistic} and its associated \code{p-value}.
If \code{x} and \code{y} are either matrices or dataframes, a \code{statistic} and \code{p-value} will be returned
for each row.
}
\description{
Peforms the test for two multinomial vectors
testing \eqn{H_o:} the underlying multinomial probability vectors are the same
vs. \eqn{H_1:} they are different.
}
\examples{
#Generate two vectors from the same distributions:
data <- genMultinomialData(sample_size=1)

#Perform test:
multinom.test(x=data[[1]],y=data[[2]])

#Generate 1000 vectors from each of two different distributions:
data <- genMultinomialData(null_hyp=FALSE,sample_size=1000)

#Perform test (compare the ith row of x to the ith row of y for all rows):
result <- multinom.test(x=data[[1]],y=data[[2]])
#Return power of test:
mean(result$pvalue)
}
\seealso{
Amanda Plunkett & Junyong Park (2018) \emph{Two-Sample Test for Sparse High Dimensional Multinomial Distributions}, TEST, \url{https://doi.org/10.1007/s11749-018-0600-8}
}
